Move update_json method to its own file
